Initial commit
I made a bit hackier version of ast_yieldifier.py for my own use, but
ended up cleaning it up for an answer on SO. And since I was already at
it, I decided to write bytecode_yieldifier.py too, to demonstrate the
other approach to doing it. Neither is perfect, but should be good
enough for most non-crazy functions, or as a base for something more
sophisticated.